# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oculus-Feldsher API",
    "description": "Медицинская платформа для удалённой подготовки офтальмологических пациентов к операции",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Локальный сервер"
    }
  ],
...
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Регистрация нового пользователя",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email уже зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вход в систему",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Вход выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учётные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Обновление токенов",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Токены обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Недействительный refresh токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/patient-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вход пациента по коду доступа",
        "description": "Публичный endpoint для входа пациента используя 8-значный код доступа. Возвращает JWT токены для доступа к защищённым endpoints.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Вход выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Неверный код доступа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/telegram-token-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Вход через одноразовый токен из Telegram",
        "description": "Публичный endpoint для входа пациента используя одноразовый токен, полученный через Telegram бот командой /login. Токен действителен 15 минут и может быть использован только один раз.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Вход выполнен успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Недействительный или истёкший токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Получить текущего пользователя",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Текущий пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Выход из системы (отзыв refresh токена)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Выход выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Требуется авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/districts": {
      "get": {
        "tags": [
          "Districts"
        ],
        "summary": "Список районов",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список районов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseDistrictList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Districts"
        ],
        "summary": "Создать район (ADMIN)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDistrictRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Район создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseDistrict"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/districts/{id}": {
      "get": {
        "tags": [
          "Districts"
        ],
        "summary": "Получить район по ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Район",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseDistrict"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Districts"
        ],
        "summary": "Обновить район (ADMIN)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDistrictRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseDistrict"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      },
      "delete": {
        "tags": [
          "Districts"
        ],
        "summary": "Удалить район (ADMIN)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Удалено"
          },
          "403": {
            "description": "Доступ запрещён"
          },
          "404": {
            "description": "Не найдено"
          }
        }
      }
    },
    "/api/v1/patients": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Список пациентов (с учётом RBAC)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "IN_PROGRESS",
                "PENDING_REVIEW",
                "APPROVED",
                "NEEDS_CORRECTION",
                "SCHEDULED",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список пациентов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponsePatientList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Создать пациента (DISTRICT_DOCTOR/ADMIN)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пациент создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponsePatient"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          }
        }
      }
    },
    "/api/v1/patients/dashboard": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Статистика дашборда",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Количество по статусам",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/status/{code}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Публичный статус пациента (без авторизации)",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Публичный статус пациента",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponsePatientPublic"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          }
        }
      }
    },
    "/api/v1/patients/{id}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Получить пациента по ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пациент",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponsePatient"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено"
          }
        }
      },
      "patch": {
        "tags": [
          "Patients"
        ],
        "summary": "Обновить данные пациента",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлено"
          },
          "400": {
            "description": "Ошибка"
          }
        }
      }
    },
    "/api/v1/patients/{id}/status": {
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Изменить статус пациента",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Статус изменён"
          },
          "400": {
            "description": "Ошибка"
          }
        }
      }
    },
    "/api/v1/patients/{id}/batch-update": {
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Пакетное обновление пациента (для оффлайн-режима)",
        "description": "Атомарное обновление данных пациента, статуса и чек-листа с обнаружением конфликтов",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пакетное обновление выполнено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации"
          },
          "409": {
            "description": "Конфликт данных (timestamp устарел)"
          }
        }
      }
    },
    "/api/v1/checklists/patient/{patientId}": {
      "get": {
        "tags": [
          "Checklists"
        ],
        "summary": "Получить пункты чек-листа пациента",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пункты чек-листа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseChecklistItems"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checklists/patient/{patientId}/progress": {
      "get": {
        "tags": [
          "Checklists"
        ],
        "summary": "Получить прогресс выполнения чек-листа",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Прогресс",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseChecklistProgress"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checklists/{id}": {
      "patch": {
        "tags": [
          "Checklists"
        ],
        "summary": "Обновить пункт чек-листа",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChecklistItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Обновлено"
          },
          "400": {
            "description": "Ошибка"
          }
        }
      }
    },
    "/api/v1/checklists/{id}/review": {
      "post": {
        "tags": [
          "Checklists"
        ],
        "summary": "Проверка пункта хирургом (SURGEON/ADMIN)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewChecklistItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Проверено"
          },
          "400": {
            "description": "Ошибка"
          }
        }
      }
    },
...
    "/api/v1/media/patient/{patientId}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Список медиафайлов пациента",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список медиафайлов"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT токен можно передать через заголовок Authorization: Bearer <token> или через query параметр ?token=<token>"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}